Rewritten RtMidi wrapper classes
Was not happy with the way responsibility was between old manager and devices, so rewrote them to better separate responsibilities of the manager (enumerate input/output devices and provide info) and the devices themselves (responsible for opening/closing ports and holding the device handle).
Didn't re-implement any functionality related to virutal ports, as this is only supported on certain platforms using RtMidi - and I am aiming for the lowest common denominator.