Fixed managed->unmanaged issue
When passing a delegate onto unmanaged code, it is important to explicitly keep a pointer in managed code to the resulting delegate, when using syntactic sugar (just pointing to the method name) this translates into new DelegateType(MethodName), thus the delegate introduced is only passed to the unmanaged code and therefore un a future GC the delegate might be collected resulting in an exception when the unmanaged code tries to invoke the delegate.

The solution is simple, keep a pointer in managed code to the delegate, and pass the pointer to unmanaged code.

The dangers of syntactic sugar at play.